explicitly pass GridView into GuardedVisit
The GuardedVisit TMP is now explicitly templated on the GridView and
the functions within it are passed the GridView as a parameter. This
change was necessary because the GuardedVisit struct cannot access the
GridFunctionSpace embedded in the LocalFunctionSpace (it is private).
We thus have to extract it directly from the MultiDomainFunctionSpace
embedded in the MultiDomainLocalFunctionSpace.