extended signature of SubProblemLFS with SubProblem and Constraints
Unlike initially assumed, I was not able to implement
a version of constraints() which only takes the
subproblem as parameter. This is not possible, as
we cannot say whether to use the trial grid function
space or the test grid function space. Thus constraints()
now directly takes the SubProblemLocalFunctionSpace.

This change requires the SubProblemLocalFunctionSpace
to provide its associated subproblem and constraints
engine.

Added those parameters and adjusted all instantiation
sites.