BUGFIX: Pass correct implementation type to Function base class in macros
There was a bug in the macros SIMPLE_ANALYTIC_FUNCTION
and SIMPLE_BOUNDARY_FUNCTION: It always passed F
respective B as the implementation class to the base
class, irrespective of the actual class name.

This lead to very weird behavior, where within function
G the function body of function F was executed.