Revert "BUGFIX: Don't try to pass an intersection into lambda_volume!"
This is what happens if you are in a bug fixing frenzy... You just
work too fast and get sloppy. In this case, looking at the context
of the call would have helped: Turns out the problem was not passing
the wrong parameter, but calling the wrong function.

Will be fixed in next commit.

This reverts commit 2bd9c12a64d4f531c30fbd12939b9e8b70437ff6.

Conflicts:

	dune/pdelab/multidomain/multidomaingridoperatorspace.hh