Cleaned up parameter function macros
The parameter function macros were ugly, as they
took the function body as a template parameter,
causing all kinds of parsing problems. This is now
simply avoided by splitting the macro into a header
and a footer macro. This also allows specifying the
names of the function parameters.

Also removed the second boundary condition type
function and cleaned up the main() a little bit.