add new test example reactiondiffusion for coupled reaction-diffusion / diffusion problem
The new problem is a simple extension of the poisson
example. Instead of solving an identical poisson
problem on both parts of the domain, it creates a
second GridFunctionSpace on the left part and solves
the reaction-diffusion example problem from pdelab-howto
on that subdomain. On the other subdomain, the poisson
operator is replaced by the diffusion operator to
support a diffusion coefficient different from 1.
Coupling is again implicit, i.e. built into the
underlying space.