defer the setup work of the SubProblemLFS to the bind() stage
The multi-component SubProblemLFS needs to calculate
its size and copy the global indices from its
components. As this data depends on the entity, it
is impossible to do in setup(). This adds a new
method bind() (doesn't need the entity as parameter,
as it only depends on the MultiLFS), which needs
to be called every time the MultiLFS gets bound to
a different entity.