make operator_applier data modular and move class to separate header
The operator_applier and the connected TMP machinery
in multidomaingridfunctionspace.hh are also needed
for the instationary MDGOS, but the data required
by the functors invoked on the subproblems will differ
in some cases. There will especially not be local
function spaces all the time. In order to support that
scenario and to increase flexibility for future use
cases, this patch separates the invocation logic from
the data carried by operator_applier. For that purpose,
the data part of the class is split into many small
purpose-built classes, from which the user can pick
the required set and pass them as template parameters
to operator_applier. Internally, the class then simply
multiply inherits from all template parameters.

As the resulting mechanism is fairly general, it
can now be found in the new header operatorapplier.hh