add optional support for accessing MDGFS children by type
With this patch, it becomes possible to access the
children of a MDGFS by their type, provided that
all children have a distinct type (otherwise, the
lookup would not make sense as it would be ambiguous).

This precondition is verified by the MDGFS, which
will only enable the functionality if the condition
is met.