increase flexibility of child introspection in MultiDomainGFS
Up to now, the MultiDomainGridFunctionSpace could
only differentiate between grid function spaces defined
on the MultiDomainGrid and grid function spaces defined
on a SubDomainGrid. As we need greater flexibility
for handling the new CouplingGridFunctionSpace, this
patch introduces tag types for the three different
flavours of children currently supported and adds
the mechanism for handling the tagging and translating
to the old way of doing things.