parametrize the TMP for visiting children on the visitor
Up to now, the visitor was only needed for differentiating
between multidomain and subdomain GFS, and wrap the
multidomain entity in a subdomain one in the latter
case.
Now we also need to handle coupling spaces. With a
different visitor, we can just reuse the existing
TMP (with a bit of abuse...).