[SubProblem/SubProblemLocalFunctionSpace] Remove custom constraints operator
Having a custom constraints operator on the SubProblem
was not a good idea, as in the case of a composite
space for the subproblem, the whole tree has to be
assembled using a single constraints operator, which
will not work in general. So now the user has to give
the correct constraints operator to the original
GridFunctionSpaces, but it is still possible to
control their behavior using the parameter objects.