[LocalAssembler Engines] BUGFIX: Don't mix up two separate tests in a single if-else
When re-adding checks to only apply skeleton methods
two-sided when requested by the local operator in
12757951, the test was performed together with the
skeleton / boundary test, causing a boundary term
to be assembled for intersections that are in fact
skeleton intersections, but should not be assembled
because they are in inverse direction.

Fixed by separating the two tests.