[Assembler] Add general facility for guarded calls to local operators
While most of the time, the LocalOperatorCallSwitch
from standard PDELab is not required in our assembler
(calls are filtered before entering the functors),
its functionality remains necessary for things like
unified handling of boundaries and skeleton.

Unfortunately, the design of that call switch has
several drawbacks:
- verbose syntax
- lack of extensibility (it is a specialized struct)
- error-prone repetitive code

This patch introduces a new facility based on free-
standing pairs of functions that use SFINAE to detect
whether a guarded call should actually be performed.
These function are placed in the namespace
Dune::PDELab::guarded_call, which can easily be
extended without having to touch the original file.
Finally, to reduce the burden of the repetitive function
definitions, there is a simple macro that generates
the functions for a pair of (guard constant, method name).
The generated functions rely on variadic templates
and rvalue references to provide perfect forwarding
of method arguments in a generic fashion.