[VTK] Add new MultiDomain-specific add_solution_to_vtk_writer() function
The vanilla VTK output infrastructure cannot cope with a VTKWriter defined
on a subdomain, as the MultiDomainGridFunctionSpace is defined on the
MultiDomainGrid. Moreover, it does not make sense to simply output all
leaf spaces; we only want to pick those that are defined on the GridView
of the VTKWriter.

In order to work around this issue, this patch exploits the extension
mechanism introduced in dune-pdelab r2014 and r2015. The changes in
PDELab make sure that LocalFunctionSpaces defined on GridViews of
different type than the VTKWriter will be completely skipped, avoiding
compilation problems. This patch contains the other half of the puzzle
in form of an extended common data holder that will automatically
convert SubDomainGrid entities to MultiDomainGrid entities (remember,
the MultiDomainLocalFunctionSpace always lives on the MultiDomainGrid).

Apart from that, we also provide a simple predicate to only pick
spaces defined on the correct subdomain.