[VTK] Adapt MultiDomain VTK output support to new name generator component introduced in dune-pdelab r2057
This patch switches the MultiDomain version of add_solution_to_vtk_writer to work with
a name generator instead of a base name, as required by the changed PDELab API. Note
that in constrast to the vanilla PDELab version of the function, the order of the name
generator and the predicate arguments has been switched, as in a MultiDomain context,
predicates are almost always required. Having the name generator after the predicate
allows it to be defaulted and thus reduces code verbosity.