[navierstokesdarcy-analytic] Add source term to interface mass conservation equation
Compared to Ida's original example, we cannot control the relationship between Darcy
pressure and Darcy velocity, as we do not track the velocity in a variable. As the actual
velocity of the analytical example is not equal to the gradient of the pressure along the
interface, this causes a discontinuity of the velocity across the interface that has to be
compensated with a source term.

This patch adds a new source term, h3, to the mass conservation equation of the coupling
operator and provides the correct analytical source term for the example.