[StokesDarcyCoupling] Redesign handling of stress term in coupling operator
The Beavers-Joseph-Saffmann conditions impose Neumann-Type boundary conditions
on both the Darcy and the free-flow domains along their common interface. In the case
of the free-flow domain, this boundary condition is vector-valued, as it applies to the
normal part of the viscuous stress tensor. The second and third equations of the interface
conditions prescribe values for the normal and the tangential part of this vector, respectively
(relative to the interface orientation at the current position).

Unfortunately, most formulations of the conditions treat both parts as scalar equations,
but while this is absolutely correct for the normal part (the orientation of the normal is
always well-defined), doing so for the tangential part either requires a lot of care (in 2D) or
is outright impossible (in 3D), as there is no unique tangential direction, making it impossible
to reduce the tangential interface condition to a scalar equation.

This patch takes the vector-valued nature of the third condition into account and assembles the
BJS conditions accordingly. Note that this also affects the formulation of the underlying equations:
The forcing term for the tangential part of the stress tensor is now a vector-valued quantity, as
we cannot in general assume that its orientation will coincide with the tangential flow direction
on the interface.