[MultiDomainLocalFunctionSpace] Avoid resizing DOFIndex storage vector on every bind()
We already set the vector to the maximum possible size (as returned by maxLocalSize() )
at construction time, making the resize operation at bind() time rather redundant...

This won't make much of a difference for fixed-size spaces, but avoids constructor
and destructor calls when the local size of a space is not constant.

This patch mirrors the change in PDELab r2081.