[DataHandle] Implement data handle support for MultiDomainGridFunctionSpace
This patch adds data handle support based on the new backend API to MDGFS. Unfortunately,
I didn't figure out how to do this as smoothly as in the standard PDELab case, where we
don't even need the GridFunctionSpace for this anymore, but can operate exclusively on
the ordering. The problem with that approach is that in the MDGFS, I need to convert between
the MultiDomainGrid and possible SubDomainGrids, which requires access to the GridView and
thus to the GridFunctionSpace.

So I unfortunately had to copy over the DataHandleProvider from vanilla PDELab and adapt
it to the multidomain scenario. Still, the implementation tries to dispatch to the vanilla
version as early as possible by only handling the direct child level of the MDGFS. Differentation
between the different types of child spaces is based on tag dispatch on the flavor tag as usual.