[Documentation] Add some decent documentation
Add a README, some pointers for building and hacking on the lib and a
file with release notes. The files are all formatted for Markdown to
enable nicely looking pages on GitHub.

The original README has been kept for compatibility with "make dist",
but is now just a symlink to README.md.