[CouplingLFS] Make data members public so they can be accessed by the PDELab visitors
Most of the actual funtionality of the local function space is
implemented in a couple of TypeTree visitors that we reuse unchanged
from PDELab. Those visitors need access to the internals of the LFS, but
unfortunately, they live in an anonymous namespace, which makes it
impossible to declare them as friends outside of their own source file.

So we just take the dirty way out and make everything in CouplingLFS
public; if anyone is stupid enough to play around with those members,
they deserve what's coming for them.