[PowerCouplingGFS] Update to work with improved PowerCompositeGridFunctionSpaceBase
The PowerCompositeGridFunctionSpaceBase mixin now expects both the
ordering and the backend tags in its template parameter list and also
requires the derived class to pass instances of both types in the
constructor. So this patch changes all constructor signatures of
PowerCouplingGFS to accept instances of those two types. Apart from
that, it also fixes the interface to the Ordering to do the whole
create-on-demand thing.